+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/block/fluid/BlockFluidGloamberryWine.java
@@ -1,0 +1,14 @@
+package com.codetaylor.mc.pyrotech.modules.core.block.fluid;
+
+import net.minecraft.block.material.Material;
+import net.minecraftforge.fluids.BlockFluidClassic;
+import net.minecraftforge.fluids.Fluid;
+
+public class BlockFluidGloamberryWine
+    extends BlockFluidClassic {
+
+  public BlockFluidGloamberryWine(Fluid fluid) {
+
+    super(fluid, Material.WATER);
+  }
+}